+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/ui/channels/list/ChannelList.kt
@@ -1,10 +1,13 @@
 package io.getstream.chat.android.compose.ui.channels.list
 
 import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.BoxScope
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.lazy.LazyListState
+import androidx.compose.foundation.lazy.rememberLazyListState
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.painterResource
@@ -36,12 +39,14 @@ import io.getstream.chat.android.offline.ChatDomain
  * @param modifier Modifier for styling.
  * @param viewModel The ViewModel that loads all the data and connects it to the UI. We provide a
  * factory that builds the default ViewModel in case the user doesn't want to provide their own.
+ * @param lazyListState State of the lazy list that represents the list of channels. Useful for controlling the scroll state.
  * @param onLastItemReached Handler for pagination, when the user reaches the last item in the list.
  * @param onChannelClick Handler for a single item tap.
  * @param onChannelLongClick Handler for a long item tap.
  * @param loadingContent Composable that represents the loading content, when we're loading the initial data.
  * @param emptyContent Composable that represents the empty content if there are no channels.
  * @param emptySearchContent Composable that represents the empty content if there are no channels matching the search query.
+ * @param helperContent Composable that represents the helper content. Empty by default, but can be used to implement scroll to top button.
  * @param itemContent Composable that allows the user to completely customize the item UI.
  * It shows [ChannelItem] if left unchanged, with the actions provided by [onChannelClick] and
  * [onChannelLongClick].
@@ -62,6 +67,7 @@ public fun ChannelList(
             )
         )
     ),
+    lazyListState: LazyListState = rememberLazyListState(),
     onLastItemReached: () -> Unit = { viewModel.loadMore() },
     onChannelClick: (Channel) -> Unit = {},
     onChannelLongClick: (Channel) -> Unit = { viewModel.selectChannel(it) },
@@ -73,6 +79,7 @@ public fun ChannelList(
             modifier = modifier
         )
     },
+    helperContent: @Composable BoxScope.() -> Unit = {},
     itemContent: @Composable (ChannelItemState) -> Unit = { channelItem ->
         DefaultChannelItem(
             channelItem = channelItem,
@@ -87,12 +94,14 @@ public fun ChannelList(
         modifier = modifier,
         channelsState = viewModel.channelsState,
         currentUser = viewModel.user.value,
+        lazyListState = lazyListState,
         onLastItemReached = onLastItemReached,
         onChannelClick = onChannelClick,
         onChannelLongClick = onChannelLongClick,
         loadingContent = loadingContent,
         emptyContent = emptyContent,
         emptySearchContent = emptySearchContent,
+        helperContent = helperContent,
         itemContent = itemContent,
         divider = divider
     )
@@ -113,12 +122,14 @@ public fun ChannelList(
  * @param channelsState Current state of the Channel list, represented by [ChannelsState].
  * @param currentUser The data of the current user, used various states.
  * @param modifier Modifier for styling.
+ * @param lazyListState State of the lazy list that represents the list of channels. Useful for controlling the scroll state.
  * @param onLastItemReached Handler for pagination, when the user reaches the end of the list.
  * @param onChannelClick Handler for a single item tap.
  * @param onChannelLongClick Handler for a long item tap.
  * @param loadingContent Composable that represents the loading content, when we're loading the initial data.
  * @param emptyContent Composable that represents the empty content if there are no channels.
  * @param emptySearchContent Composable that represents the empty content if there are no channels matching the search query.
+ * @param helperContent Composable that represents the helper content. Empty by default, but can be used to implement scroll to top button.
  * @param itemContent Composable that allows the user to completely customize the item UI.
  * It shows [ChannelItem] if left unchanged, with the actions provided by [onChannelClick] and
  * [onChannelLongClick].
@@ -129,6 +140,7 @@ public fun ChannelList(
     channelsState: ChannelsState,
     currentUser: User?,
     modifier: Modifier = Modifier,
+    lazyListState: LazyListState = rememberLazyListState(),
     onLastItemReached: () -> Unit = {},
     onChannelClick: (Channel) -> Unit = {},
     onChannelLongClick: (Channel) -> Unit = {},
@@ -140,6 +152,7 @@ public fun ChannelList(
             modifier = modifier
         )
     },
+    helperContent: @Composable BoxScope.() -> Unit = {},
     itemContent: @Composable (ChannelItemState) -> Unit = { channelItem ->
         DefaultChannelItem(
             channelItem = channelItem,
@@ -157,7 +170,9 @@ public fun ChannelList(
         !isLoading && channels.isNotEmpty() -> Channels(
             modifier = modifier,
             channelsState = channelsState,
+            lazyListState = lazyListState,
             onLastItemReached = onLastItemReached,
+            helperContent = helperContent,
             itemContent = itemContent,
             divider = divider
         )