+++ CHANGELOG.md
@@ -87,6 +87,8 @@ Consider migrating to `stream-chat-android-ui-components` or `stream-chat-androi
 ### ✅ Added
 - Added [`Custom Attachments guide`](https://getstream.io/chat/docs/sdk/android/composee/guides/adding-custom-attachments/). [#2967](https://github.com/GetStream/stream-chat-android/pull/2967)
 - Added `onHeaderAvatarClick` parameter to the `ChannelsScreen` component. [#3016](https://github.com/GetStream/stream-chat-android/pull/3016)
+- Exposed `lazyListState` for the `ChannelList` and `Channels` components that allows you to control the scrolling behavior and state. [#3049](https://github.com/GetStream/stream-chat-android/pull/3049)
+- Exposed `helperContent` for the `ChannelList` and `Channels` components that allows you to implement a helper UI such as scroll to top button for the channel list. [#3049](https://github.com/GetStream/stream-chat-android/pull/3049)
 
 ### ⚠️ Changed
 - Renamed `onHeaderClickAction` parameter to `onHeaderActionClick` for the `ChannelsScreen` component. [#3016](https://github.com/GetStream/stream-chat-android/pull/3016)